pyproject.toml
[project]
name = "memory-price-tracker"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "selectolax>=0.3",
    "gspread>=6.0",
    "google-auth>=2.0",
    "python-dotenv>=1.0",
    "playwright>=1.44",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[project.scripts]
mpt = "src.main:main"
